pkg/bisect: don't depend on res.rep for skipped commits
res.rep is not populated for build/boot failures. I can only be
trusted when the bisection verdict was 'bad', so that's what env.test
callers should always test for.

Tracing res.rep through the code is also not easy for a human. It's
regularly overwritten, but since it's a pointer one has to dig deep
into the call stack to find out whether the asigned value could
possibly be null.